# Performance(diff): faster rename detection