Fixes iterative/dvc#10515

## Notes and questions
- I presume the reason for sorting `added` and `deleted` before was to optimize the runtime (more likely to get an early hit in the inner loop). I included this now maintain the former output. In the case of multiple files with the same hash, changing this may result in other pairs being detected as renames. However, if maintaining this is not a priority, I think we can remove it, making it simpler. It will still be deterministic.
- I used asserts for verifying `change.new` and `change.old`, as this seemed to be the convention. LMK. if I should change for exceptions, or if we should just assume `_diff` to always supply valid changes.
- I'm only just aware of the https://github.com/iterative/dvc-bench repo. Should we add a test case to it that would have uncovered iterative/dvc#10515?
- It is not entirely clear to me why `HashInfo` has `unsafe_hash=True`, however, as far as I understant, it should not affect this implementation.

## Benchmark

Consider a similar setup as the minimal reproduction in iterative/dvc#10515.
We make three versions of our dataset:
- **Tag1**: `mkdir data; for i in {1..10000}; do echo $i > data/file_$i; done`
- **Tag2**: `mkdir data; for i in {1..10000}; do echo $i > data/file_$i.ext; done`
- **Tag3**: `mkdir data; for i in {1..10000}; do echo $((i + 20000)) > data/file_$i.ext; done`
From Tag1 -> Tag2, we rename the files, without changing the content. From Tag2-> Tag3, we change the content of the files. 

Run `dvc diff <from> <to>`.

| Change | Old | New |
| --- | --- | --- |
| Tag1 -> Tag2 | 0.49 s | 0.50 s | 
| Tag2 -> Tag3 | 0.46 s |  0.46 s |
| Tag1 -> Tag3 | 14.6 s | 0.56 s |

Ps. Results show timings from only one experiment.