@Northo 2024-08-09
> @Northo, can you share benchmarks or profiling information? Also, can you please add a description?

@skshetry, sorry. Inteded to make a draft pull request only, at this time. Will supply more detailed inormation soon. For now, see iterative/dvc#10515.