@Northo 2024-08-12
> > @skshetry, one benchmark test failed with regression. I cannot see that it is related to these changes, and also see that it first failed and then suceeded on #552. Have I made a mistake, or is it just flaky?
> 
> yeah, it's flaky. Please ignore.

Ok, thanks! In that case, consider my PR open for review:)