@skshetry 2024-08-13
I tried your script above (but created 100,000 files instead of 10,000). But I did not see any meaningful changes, and in fact it was slightly slower than before.


Since we are using an iterator, I used the following patch to isolate `_detect_renames`.

```patch
diff --git a/src/dvc_data/index/diff.py b/src/dvc_data/index/diff.py
index 40fd5eb..491dfcf 100644
--- a/src/dvc_data/index/diff.py
+++ b/src/dvc_data/index/diff.py
@@ -5,6 +5,7 @@ from typing import TYPE_CHECKING, Any, Callable, Optional, cast

 from attrs import define
 from fsspec.callbacks import DEFAULT_CALLBACK, Callback
+import funcy

 if TYPE_CHECKING:
     from dvc_data.hashfile.hash_info import HashInfo
@@ -328,6 +329,9 @@ def diff(  # noqa: PLR0913

     if with_renames and old is not None and new is not None:
         assert not meta_only
-        yield from _detect_renames(changes)
+        changes_list = list(changes)
+        with funcy.print_durations("Detecting renames"):
+            changes = list(_detect_renames(changes_list))
+        yield from changes
     else:
         yield from changes
```