@Northo 2024-08-13
> EDIT: Sorry, I was not deleting the directories between each stages. I see the improvements now. :)

Happy to hear that! I was struggling trying to figure out why it could not be reproduced;) 
Thanks for the well-formed reproduction script!