@Northo 2024-08-13
> Thank you for contributing, and making an amazing improvement in performance to dvc. 🙂

Thanks for really swift follow-up on the PR @skshetry !